refactor / get all events with organizers
